# Datum reader improvements